Created a save file for selected directory and warning pop-up
Made the selected directory persistent after first use, meaning the when
the user selects an output directory it is saved and pre-selected for
the next time. Added a warning pop-up if the user does not input a
complete web site address. Also added Javadoc comments for all classes
and methods.

Change-Id: I8e7d4ccda371500d4714eb918920890318adaa7b